[Refactor] Move http_client to HttpCore
