+++ be/CMakeLists.txt
@@ -544,7 +544,6 @@ set(STARROCKS_LINK_LIBS
     ${WL_WHOLE_ARCHIVE_PREFIX} Exprs ${WL_WHOLE_ARCHIVE_SUFFIX}
     FileSystem
     Formats
-    HttpClient
     IO
     Storage
     Rowset
