+++ be/src/common/utils.h
@@ -16,37 +16,3 @@
 // under the License.
 
 #pragma once
-
-#include <string>
-
-namespace starrocks {
-
-struct AuthInfo {
-    std::string user;
-    std::string passwd;
-    std::string cluster;
-    std::string user_ip;
-    // -1 as unset
-    int64_t auth_code = -1;
-};
-
-template <class T>
-void set_request_auth(T* req, const AuthInfo& auth) {
-    if (auth.auth_code != -1) {
-        // If 'auth_code' is set, no need to set other info.
-        req->__set_auth_code(auth.auth_code);
-        // User name and passwd is unused, but they are required field.
-        // so they have to be set.
-        req->user = "";
-        req->passwd = "";
-    } else {
-        req->user = auth.user;
-        req->passwd = auth.passwd;
-        if (!auth.cluster.empty()) {
-            req->__set_cluster(auth.cluster);
-        }
-        req->__set_user_ip(auth.user_ip);
-    }
-}
-
-} // namespace starrocks
