+++ be/src/http/action/transaction_stream_load.cpp
@@ -27,6 +27,7 @@
 #include <rapidjson/prettywriter.h>
 #include <thrift/protocol/TDebugProtocol.h>
 
+#include "base/auth/auth_info.h"
 #include "base/testutil/sync_point.h"
 #include "base/time/time.h"
 #include "base/uid_util.h"
@@ -36,7 +37,6 @@
 #include "common/logging.h"
 #include "common/system/master_info.h"
 #include "common/util/debug_util.h"
-#include "common/utils.h"
 #include "gen_cpp/FrontendService.h"
 #include "gen_cpp/FrontendService_types.h"
 #include "gen_cpp/HeartbeatService_types.h"
