+++ be/src/runtime/batch_write/isomorphic_batch_write.cpp
@@ -25,7 +25,6 @@
 #include "common/config_merge_commit_fwd.h"
 #include "common/system/master_info.h"
 #include "common/util/bthreads/executor.h"
-#include "common/utils.h"
 #include "gen_cpp/FrontendService.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "http/http_common.h"
