+++ be/src/runtime/stream_load/stream_load_executor.cpp
@@ -38,6 +38,7 @@
 
 #include <string_view>
 
+#include "base/auth/auth_info.h"
 #include "base/testutil/sync_point.h"
 #include "base/utility/defer_op.h"
 #include "common/config_ingest_fwd.h"
@@ -45,7 +46,6 @@
 #include "common/status.h"
 #include "common/statusor.h"
 #include "common/system/master_info.h"
-#include "common/utils.h"
 #include "gen_cpp/FrontendService.h"
 #include "runtime/client_cache.h"
 #include "runtime/exec_env.h"
