+++ be/test/http/CMakeLists.txt
@@ -14,6 +14,7 @@
 
 set(HTTP_CORE_TEST_FILES
     ../base/cxa_throw_wrap.cpp
+    http_client_test.cpp
     http_core_test.cpp
     http_metrics_test.cpp
 )
