+++ be/test/http/http_client_test.cpp
@@ -21,7 +21,7 @@
 
 #include "base/testutil/assert.h"
 #include "boost/algorithm/string.hpp"
-#include "common/config_storage_fwd.h"
+#include "common/config_ingest_fwd.h"
 #include "common/logging.h"
 #include "http/ev_http_server.h"
 #include "http/http_auth.h"
@@ -71,6 +71,11 @@ public:
     }
 };
 
+class HttpClientTestNotFoundHandler : public HttpHandler {
+public:
+    void handle(HttpRequest* req) override { HttpChannel::send_reply(req, HttpStatus::NOT_FOUND, "Not Found"); }
+};
+
 class HttpClientTestHeaderHandler : public HttpHandler {
 public:
     void handle(HttpRequest* req) override {
@@ -100,6 +105,7 @@ public:
 
 static HttpClientTestSimpleGetHandler s_simple_get_handler = HttpClientTestSimpleGetHandler();
 static HttpClientTestSimplePostHandler s_simple_post_handler = HttpClientTestSimplePostHandler();
+static HttpClientTestNotFoundHandler s_not_found_handler = HttpClientTestNotFoundHandler();
 static HttpClientTestHeaderHandler s_header_handler = HttpClientTestHeaderHandler();
 static HttpClientTestMultiHeaderHandler s_multi_header_handler = HttpClientTestMultiHeaderHandler();
 
@@ -113,10 +119,12 @@ public:
     ~HttpClientTest() override = default;
 
     static void SetUpTestCase() {
+        config::streaming_load_max_mb = 102400;
         s_server = new EvHttpServer(0);
         s_server->register_handler(GET, "/simple_get", &s_simple_get_handler);
         s_server->register_handler(HEAD, "/simple_get", &s_simple_get_handler);
         s_server->register_handler(POST, "/simple_post", &s_simple_post_handler);
+        s_server->register_handler(POST, "/simple_post_failed", &s_not_found_handler);
         s_server->register_handler(GET, "/header_test", &s_header_handler);
         s_server->register_handler(GET, "/multi_header_test", &s_multi_header_handler);
         ASSERT_OK(s_server->start());
@@ -240,22 +248,6 @@ TEST_F(HttpClientTest, header_override) {
     ASSERT_EQ("H1:overridden;H2:initial;H3:new_value", response);
 }
 
-TEST_F(HttpClientTest, download) {
-    HttpClient client;
-    auto st = client.init(hostname + "/simple_get");
-    ASSERT_TRUE(st.ok());
-    client.set_basic_auth("test1", "");
-    std::string local_file = ".http_client_test.dat";
-    auto st_or = client.download(local_file);
-    ASSERT_TRUE(st_or.ok());
-    char buf[50];
-    auto fp = fopen(local_file.c_str(), "r");
-    auto size = fread(buf, 1, 50, fp);
-    buf[size] = 0;
-    ASSERT_STREQ("test1", buf);
-    unlink(local_file.c_str());
-}
-
 TEST_F(HttpClientTest, download_to_memory) {
     HttpClient client;
     auto st = client.init(hostname + "/simple_get");
@@ -267,8 +259,7 @@ TEST_F(HttpClientTest, download_to_memory) {
                 value.append((const char*)data, length);
                 return Status::OK();
             },
-            config::replication_min_speed_limit_kbps, config::replication_min_speed_time_seconds,
-            config::replication_max_speed_limit_kbps);
+            50, 300, 50000);
     ASSERT_TRUE(st.ok());
     ASSERT_EQ("test1", value);
 }
@@ -300,7 +291,7 @@ TEST_F(HttpClientTest, post_normal) {
 
 TEST_F(HttpClientTest, post_failed) {
     HttpClient client;
-    auto st = client.init(hostname + "/simple_pos");
+    auto st = client.init(hostname + "/simple_post_failed");
     ASSERT_TRUE(st.ok());
     client.set_method(POST);
     client.set_basic_auth("test1", "");
